🐰 [FIX] move shared pydantic model to shared
